requirements.txt
base45
pyzbar
pillow
flynn
